+++ splashsurf/src/lib.rs
@@ -5,7 +5,5 @@ mod reconstruction;
 #[macro_use]
 mod allocator;
 mod logging;
-#[cfg(test)]
-mod tests;
 
 pub(crate) use register_counting_allocator;
