+++ splashsurf/src/main.rs
@@ -5,8 +5,6 @@ mod reconstruction;
 #[macro_use]
 mod allocator;
 mod logging;
-#[cfg(test)]
-mod tests;
 
 pub(crate) use register_counting_allocator;
